orders/src/models/ticket.ts
import mongoose from "mongoose";
//  Manages the version number of records
//  and customises the update operation to find the correct version
import { updateIfCurrentPlugin } from "mongoose-update-if-current";
import { Order, OrderStatus } from "./order";

interface TicketAttrs {
  id: string;
  title: string;
  price: number;
}

export interface TicketDoc extends mongoose.Document {
  title: string;
  price: number;
  version: number;
  isReserved(): Promise<boolean>;
}

interface TicketModel extends mongoose.Model<TicketDoc> {
  build(attrs: TicketAttrs): TicketDoc;
  findByEvent(event: {
    id: string;
    version: number;
  }): Promise<TicketDoc | null>;
}

const ticketSchema = new mongoose.Schema(
  {
    title: {
      type: String,
      required: true,
    },
    price: {
      type: Number,
      required: true,
      min: 0,
    },
  },
  {
    toJSON: {
      transform(doc, ret) {
        ret.id = ret._id;
        delete ret._id;
      },
    },
  }
);

//  Optimistic concurrency (use versioning on records)
//  By default versioning key/property is __v
ticketSchema.set("versionKey", "version");
ticketSchema.plugin(updateIfCurrentPlugin);

//  If were not using the update if current plugin we could achieve the same
//  behaviour as below
// ticketSchema.pre("save", function (done) {
//   //  @ts-ignore
//   this.$where = {
//     version: this.get("version") - 1,
//   };
// });

ticketSchema.statics.build = (attrs: TicketAttrs) => {
  return new Ticket({
    _id: attrs.id,
    title: attrs.title,
    price: attrs.price,
  });
};

ticketSchema.statics.findByEvent = (event: { id: string; version: number }) => {
  return Ticket.findOne({
    _id: event.id,
    version: event.version - 1, //  Need to find previous version (so that no events/updates have been missed)
  });
};

//  Cannot be an arrow function as we are using
//  this (this is different when using an arrow function)
ticketSchema.methods.isReserved = async function () {
  //  Run query to look at all orders.  Find an order where ticket is the one
  //  found above and order status is *not* cancelled.
  //  If found ticket is already reserved
  const existingOrder = await Order.findOne({
    ticket: this,
    status: {
      $in: [
        OrderStatus.Created,
        OrderStatus.AwaitingPayment,
        OrderStatus.Complete,
      ],
    },
  });

  return !!existingOrder;
};

const Ticket = mongoose.model<TicketDoc, TicketModel>("Ticket", ticketSchema);

export { Ticket };
